ctap: widen the hmac-secret seen-key mask too
The hmac-secret extension input parser keeps its own seen-key mask and calls
zf_ctap_mark_seen_key(); it was missed when the helper moved to 32 bits and
broke the build with -Werror=incompatible-pointer-types.

+++ src/ctap/extensions/hmac_secret.c
@@ -32,7 +32,7 @@ static uint8_t zf_hmac_secret_finish(ZfHmacSecretScratch *scratch, uint8_t statu
 uint8_t zf_ctap_hmac_secret_parse_get_assertion_input(ZfCborCursor *cursor,
                                                       ZfGetAssertionRequest *request) {
     size_t pairs = 0;
-    uint16_t seen_keys = 0;
+    uint32_t seen_keys = 0;
     bool saw_key_agreement = false;
     bool saw_salt_enc = false;
     bool saw_salt_auth = false;
